refactor: Move all public application APIs to include/app/
Move all public API headers to include/app. This means
all platform-agnostic application APIs should be accessible in
form of:
'#include <app/api.h>'

This is nice to distinguish public and private headers, as well
as platform and application functions.
