+++ app/src/imu_zephyr.c
@@ -2,8 +2,7 @@
 #include <zephyr/device.h>
 #include <zephyr/drivers/sensor.h>
 #include <zephyr/logging/log.h>
-
-#include "imu.h"
+#include <app/imu.h>
 
 LOG_MODULE_REGISTER(imu_sensor, CONFIG_SENSOR_LOG_LEVEL);
